Add `-pthread` flag to C++ builds
I'm not 100% sure when/why we do/don't need the `-pthread` flag, but the code didn't compile on the VM used for the testing, and I'm guessing that removing that flag is the problem. You don't seem to need it in our lab, though.

Maybe the issue is different C++ compilers, i.e., GNU vs CLang?